fix OS command injection in sandcat_gui_api (#74)
* fix OS command injection in sandcat_gui_api

* Update sand_gui_api.py

* Replace os.system with subprocess.call for OS command injection
* Replace regex with urllib.parse for url valiation


Former-commit-id: c84c666f083bea5c193e649a6fe8cd5bbe048d62

+++ app/sand_gui_api.py
@@ -1,9 +1,10 @@
 import os
 import shutil
+import subprocess
 
+from urllib.parse import urlparse
 from aiohttp import web
 from aiohttp_jinja2 import template
-
 
 class SandGuiApi:
 
@@ -24,11 +25,20 @@ class SandGuiApi:
     async def clone_new_site(self, request):
         await self.auth_svc.check_permissions(request)
         url = request.rel_url.query['url']
-        location = 'plugins/sandcat/static/malicious/'
-        shutil.rmtree(location, ignore_errors=True)
-        os.system("""wget -U 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/76.0.3809.100 Safari/537.36' -E -H -k -K -p -q -nH --cut-dirs=1 %s --directory %s --no-check-certificate""" % (url, location))
-        self.auth_svc.prepend_to_file('%s/index.html' % location, '<script src="/sandcat/js/malicious.js"></script>')
-        self.auth_svc.prepend_to_file('%s/index.html' % location, '<meta http-equiv="Expires" content="0">')
-        self.auth_svc.prepend_to_file('%s/index.html' % location, '<meta http-equiv="Pragma" content="no-cache">')
-        self.auth_svc.prepend_to_file('%s/index.html' % location, '<meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate">')
+        if self._uri_validator(url):
+            location = 'plugins/sandcat/static/malicious/'
+            shutil.rmtree(location, ignore_errors=True)
+            user_agent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/76.0.3809.100 Safari/537.36"
+            subprocess.call(["wget","-U", user_agent, "-E","-H","-k","-K","-p","-q","-nH","--cut-dirs=1", url, "--directory", location, "--no-check-certificate"], shell=False)
+            self.auth_svc.prepend_to_file('%s/index.html' % location, '<script src="/sandcat/js/malicious.js"></script>')
+            self.auth_svc.prepend_to_file('%s/index.html' % location, '<meta http-equiv="Expires" content="0">')
+            self.auth_svc.prepend_to_file('%s/index.html' % location, '<meta http-equiv="Pragma" content="no-cache">')
+            self.auth_svc.prepend_to_file('%s/index.html' % location, '<meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate">')
         return web.Response()
+
+    def _uri_validator(self, url):
+        try:
+            result = urlparse(url)
+            return all([result.scheme, result.netloc, result.path])
+        except:
+            return False
